mouth storage fixes from #2496
