+++ Content.Shared/DeltaV/Storage/EntitySystems/SharedMouthStorageSystem.cs
@@ -4,6 +4,7 @@ using Content.Shared.Damage;
 using Content.Shared.DeltaV.Storage.Components;
 using Content.Shared.Examine;
 using Content.Shared.IdentityManagement;
+using Content.Shared.Nutrition;
 using Content.Shared.Standing;
 using Content.Shared.Storage;
 using Content.Shared.Storage.EntitySystems;
@@ -27,6 +28,7 @@ public abstract class SharedMouthStorageSystem : EntitySystem
         SubscribeLocalEvent<MouthStorageComponent, DisarmedEvent>(DropAllContents);
         SubscribeLocalEvent<MouthStorageComponent, DamageChangedEvent>(OnDamageModified);
         SubscribeLocalEvent<MouthStorageComponent, ExaminedEvent>(OnExamined);
+        SubscribeLocalEvent<MouthStorageComponent, IngestionAttemptEvent>(OnIngestAttempt);
     }
 
     protected bool IsMouthBlocked(MouthStorageComponent component)
@@ -80,5 +82,19 @@ public abstract class SharedMouthStorageSystem : EntitySystem
             var subject = Identity.Entity(uid, EntityManager);
             args.PushMarkup(Loc.GetString("mouth-storage-examine-condition-occupied", ("entity", subject)));
         }
+    }
+
+    // Attempting to eat or drink anything with items in your mouth won't work
+    private void OnIngestAttempt(EntityUid uid, MouthStorageComponent component, IngestionAttemptEvent args)
+    {
+        if (!IsMouthBlocked(component))
+            return;
+
+        if (!TryComp<StorageComponent>(component.MouthId, out var storage))
+            return;
+
+        var firstItem = storage.Container.ContainedEntities[0];
+        args.Blocker = firstItem;
+        args.Cancel();
     }
 }
