+++ Resources/Prototypes/DeltaV/Entities/Mobs/Species/rodentia.yml
@@ -9,7 +9,7 @@
     species: Rodentia
   - type: Hunger
     baseDecayRate: 0.0467 # 33% faster than usual
-  - type: Carriable # Carrying system from nyanotrasen.
+##  - type: Carriable # Carrying system from nyanotrasen. (commented out until carrying system is implemented)
   - type: Inventory
     speciesId: rodentia
   - type: Icon
@@ -100,7 +100,7 @@
       Unsexed: MaleRodentia
   - type: Rummager
   - type: AlwaysTriggerMousetrap
-  - type: PseudoItem
+##  - type: PseudoItem (commented out until carrying system is implemented)
     storedOffset: -10,0
     shape:
       - 0, 0, 3, 2
